backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
cachetools==5.5.2
Django==5.2
django-rql==4.4.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
lark-parser==0.11.0
lib-rql==2.0.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
